#612 - moved filters/helpers/mixins for appropriate directories, changed imports for these methods
