+++ src/helpers/index.js
@@ -1,5 +1,3 @@
-import he from 'he'
-
 /**
  * Create slugify -> "create-slugify" permalink  of text
  * @param {String} text
@@ -13,23 +11,7 @@ export function slugify (text) {
 }
 
 /**
- * Converts number to price string
- * @param {Number} value
- */
-export function price (value) {
-  return '$' + parseFloat(value).toFixed(2)
-}
-
-/**
- * Decodes any named and numerical character references in text
- * @param {String} value
- */
-export function htmlDecode (value) {
-  return value ? he.decode(value) : ''
-}
-
-/**
- * Re-fromat category path to be suitable for breadcrumb
+ * Re-format category path to be suitable for breadcrumb
  * @param {Array} categoryPath
  */
 export function breadCrumbRoutes (categoryPath) {
