+++ src/mixins/thumbnail.js
