+++ src/store/modules/category.js
@@ -1,6 +1,6 @@
 import * as types from '../mutation-types'
 import { entityKeyName } from '../../lib/entities'
-import { slugify } from '../../lib/filters'
+import { slugify } from '../../helpers'
 const bodybuilder = require('bodybuilder')
 import { quickSearchByQuery } from '../../api/search'
